requirements.txt
altgraph==0.17
auto-py-to-exe==2.9.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
idna==2.10
pefile==2021.5.24
Pillow==8.3.0
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pywin32-ctypes==0.2.0
requests==2.25.1
urllib3==1.26.6
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
